corner升级到2.2,hibernate 升级到3.2.0.cr4,annotations升级到3.2.0.cr1 采用了ivy+ant的自动构建方式。
git-svn-id: http://dev.bjmaxinfo.com/svn/corner/trunk/corner@1830 bda6c913-8cf9-0310-89cc-b5856a464ced

+++ src/main/java/corner/orm/tapestry/table/RelativePersistentBasicTableModel.java
@@ -106,7 +106,9 @@ public class RelativePersistentBasicTableModel<T> implements IBasicTableModel {
 			if (c == null) {
 				return 0;
 			}
-			rows = ((Integer) ((HibernateObjectRelativeUtils) this.entityService
+			
+			//TODO 考虑兼容以前的版本 Integer变为 Long
+			rows = ((Long) ((HibernateObjectRelativeUtils) this.entityService
 					.getObjectRelativeUtils()).getHibernateTemplate().execute(
 					new HibernateCallback() {
